# 1.api-gateway/src/docs/openapi/v1/combined.yaml
openapi: 3.0.3
info:
  title: eServices API Gateway (combined)
  version: 1.0.0
  description: Aggregated API for all microservices
servers:
  - url: /
    description: API Gateway
paths:
...
  /v1/auth/login:
    post:
      summary: User login
      operationId: loginUser
      tags:
        - Auth
        - Local
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Login with either email + password or username + password
              oneOf:
                - type: object
                  description: Login via email
                  required:
                    - email
                    - password
                  properties:
                    email:
                      type: string
                      format: email
                      example: user@example.com
                    password:
                      type: string
                      format: password
                      example: secret123
                - type: object
                  description: Login via username
                  required:
                    - username
                    - password
                  properties:
                    username:
                      type: string
                      example: user123
                    password:
                      type: string
                      format: password
                      example: secret123
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    example: eyJo...
                  refreshToken:
                    type: string
                    example: ooiu...
                  authId:
                    type: string
                    example: ou9qw0r8qhf...
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid credentials
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Server error
      security: []
...
  /v1/auth/register:
    post:
      summary: User registration
      operationId: registerUser
      tags:
        - Auth
        - Local
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - username
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                username:
                  type: string
                  example: user123
                password:
                  type: string
                  format: password
                  example: secret123
      responses:
        '201':
          description: Successful signup
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    example: d290f1ee-6c54-4b01-90e6-d701748f0851
                  email:
                    type: string
                    format: email
                    example: user@example.com
                  username:
                    type: string
                    example: user123
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid input
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Server error
      security: []
...
  /v1/auth/refresh-token:
    post:
      tags:
        - Auth
        - Token
      summary: Refresh an expired access token
      operationId: refreshToken
      responses:
        '200':
          description: Access token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                description: >-
                  Newly issued access token; refresh token cookie is also
                  rotated
                properties:
                  accessToken:
                    type: string
                    description: New JWT access token
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6Ikp...
                required:
                  - accessToken
        '401':
          description: No refresh token provided
          content:
            application/json:
              schema:
                type: object
                description: Standard error response
                properties:
                  success:
                    type: boolean
                    description: Indicates whether the operation failed
                    example: false
                  message:
                    type: string
                    description: Invalid or expired refresh token
                    example: Invalid or expired refresh token
                  details:
                    type: array
                    description: Optional list of error details (e.g. validation failures)
                    items:
                      type: string
                      example: Username is required
                required:
                  - success
                  - message
        '403':
          description: Malformed or tampered refresh token (signature verification failed)
          content:
            application/json:
              schema:
                type: object
                description: Standard error response
                properties:
                  success:
                    type: boolean
                    description: Indicates whether the operation failed
                    example: false
                  message:
                    type: string
                    description: Invalid or expired refresh token
                    example: Invalid or expired refresh token
                  details:
                    type: array
                    description: Optional list of error details (e.g. validation failures)
                    items:
                      type: string
                      example: Username is required
                required:
                  - success
                  - message
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                description: Standard server error response
                properties:
                  success:
                    type: boolean
                    description: Indicates the call failed
                    example: false
                  message:
                    type: string
                    description: Something went wrong
                    example: Internal server error
                  timestamp:
                    type: string
                    format: date-time
                    description: Time when the error occurred
                    example: '2025-05-23T17:00:00Z'
                  details:
                    type: array
                    description: >-
                      Optional list of additional details (e.g. stack trace
                      snippets)
                    items:
                      type: string
                      example: Database connection timed out
                required:
                  - success
                  - message
      security: []
...
  /v1/auth/otp/send:
    post:
      tags:
        - Auth
        - OTP
      summary: Send a one-time password (OTP) via email or SMS
      operationId: sendOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request to send an OTP via email or SMS
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                phone:
                  type: string
                  example: '+15555550123'
              oneOf:
                - type: object
                  description: Receive OTP via email
                  required:
                    - email
                  properties:
                    email:
                      type: string
                      format: email
                      example: user@example.com
                - type: object
                  description: Receive OTP via phone
                  required:
                    - phone
                  properties:
                    phone:
                      type: string
                      example: '+15555550123'
      responses:
        '200':
          description: OTP code sent successfully
          content:
            application/json:
              schema:
                type: object
                description: Acknowledgement that the OTP was sent
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: OTP sent via email
                required:
                  - success
                  - message
        '400':
          description: Validation failed (e.g. missing email/phone)
          content:
            application/json:
              schema:
                type: object
                description: Standard error response
                properties:
                  success:
                    type: boolean
                    description: Indicates the call failed
                    example: false
                  message:
                    type: string
                    description: Human-readable error message
                    example: Invalid or expired refresh token
                  details:
                    type: array
                    description: Additional error details or validation failures
                    items:
                      type: string
                      example: Username is required
                required:
                  - success
                  - message
        '404':
          description: User not found for given email or phone
          content:
            application/json:
              schema:
                type: object
                description: Standard error response
                properties:
                  success:
                    type: boolean
                    description: Indicates the call failed
                    example: false
                  message:
                    type: string
                    description: Human-readable error message
                    example: Invalid or expired refresh token
                  details:
                    type: array
                    description: Additional error details or validation failures
                    items:
                      type: string
                      example: Username is required
                required:
                  - success
                  - message
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                description: Internal server error response
                properties:
                  success:
                    type: boolean
                    description: Indicates the call failed
                    example: false
                  message:
                    type: string
                    description: Human-readable error message
                    example: Internal server error
                  timestamp:
                    type: string
                    format: date-time
                    description: Time when the error occurred
                    example: '2025-05-23T17:00:00Z'
                  details:
                    type: array
                    description: >-
                      Optional list of additional error details (e.g. stack
                      trace snippets)
                    items:
                      type: string
                      example: Database connection timed out
                required:
                  - success
                  - message
      security: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []